simplified mpld3 directive, made export figure into simpler png and svg buttons
