+++ client/source/js/app.js
@@ -19,7 +19,6 @@ define([
   './modules/common/normalise-height-directive',
   './modules/validations/more-than-directive',
   './modules/validations/file-required-directive',
-  './modules/mpld3-charts/chart-toolbar-directive',
   './modules/mpld3-charts/index',
   './modules/home/index',
   './modules/model/index',
@@ -53,7 +52,6 @@ define([
     'app.validations.more-than',
     'app.validations.file-required',
     'app.project',
-    'app.chart-toolbar',
     'app.ui',
     'app.ui.modal',
     'app.ui.spreadsheet-upload-hint',
