+++ client/source/js/modules/analysis/optimization.html
@@ -128,8 +128,7 @@
         </div>
         <div>
           <div ng-repeat="chart in optimizationCharts" class="chart-container __small">
-            <div mpld3-chart="chart" id="{{ chart.id }}"></div>
-            <chart-toolbar chart-type="mpld3" mpld3-chart-data="chart" chart-id="{{ chart.id }}"></chart-toolbar>
+            <div mpld3-chart="chart" chart-id="{{ chart.id }}"></div>
           </div>
         </div>
       </div>
