+++ client/source/js/modules/model/calibration.html
@@ -43,8 +43,7 @@
         <!--<export-all-data charts="calibrationChart" name="Optimization analyses" ng-if="calibrationChart.length > 0"></export-all-data>-->
       </div>
       <div ng-repeat="chart in calibrationChart" class="chart-container __small">
-        <div mpld3-chart="chart" id="{{ chart.id }}"></div>
-        <chart-toolbar chart-type="mpld3" mpld3-chart-data="chart" chart-id="{{ chart.id }}"></chart-toolbar>
+        <div mpld3-chart="chart" chart-id="{{ chart.id }} "></div>
       </div>
     </div>
     <div class="small_bordered_list" style="width: 18%;" ng-if="selectors.length > 0">
