--- client/source/js/modules/mpld3-charts/chart-toolbar.html
@@ -1,11 +1,0 @@
-<div>
-    <div ng-if="chartType === 'mpld3'">
-        <i class="fa fa-refresh chart-icon" ng-click="resetChart({$event:$event})"></i>
-        <i class="fa fa-search chart-icon" ng-click="zoomChart({$event:$event})" ng-class="{'active-icon': zoomEnabled}"></i>
-        <i class="fa fa-arrows chart-icon" ng-click="panChart({$event:$event})" ng-class="{'active-icon': panEnabled}"></i>
-    </div>
-    <div class="chart-export-buttons btn-group">
-        <button class="btn figure" ng-click="exportFigure({$event:$event})">Export figure</button>
-        <!--<button class="btn data" ng-click="exportData({$event:$event})">Export data</button>--> <!-- CK: not implemented -->
-    </div>
-</div>
