+++ client/source/js/modules/mpld3-charts/mpld3-chart.html
@@ -1,0 +1,29 @@
+<div class="mpld3-chart"></div>
+<div>
+    <div style="float:left">
+        <i class="fa fa-refresh chart-icon"
+           ng-click="resetChart({$event:$event})">
+        </i>
+        <i class="fa fa-search chart-icon"
+           ng-click="zoomChart({$event:$event})"
+           ng-class="{'active-icon': zoomEnabled}">
+        </i>
+        <i class="fa fa-arrows chart-icon"
+           ng-click="panChart({$event:$event})"
+           ng-class="{'active-icon': panEnabled}">
+        </i>
+    </div>
+    <div style="float:right">
+        <button
+            class="btn figure"
+            ng-click="exportGraphAsPng({$event:$event})">
+            PNG
+        </button>
+        <button
+            class="btn figure"
+            ng-click="exportGraphAsSvg({$event:$event})">
+            SVG
+        </button>
+        <!--<button class="btn data" ng-click="exportData({$event:$event})">Export data</button>--> <!-- CK: not implemented -->
+    </div>
+</div>
