+++ client/source/js/modules/mpld3-charts/optima-graphs.html
@@ -17,9 +17,7 @@
         <export-all-data name="Analysis scenarios" charts="graphs" ng-if="graphs"></export-all-data>
       </div>
       <div ng-repeat="chart in graphs.mpld3_graphs" ng-show="chart.isChecked()" class="chart-container __small">
-        <div mpld3-chart="chart" id="{{ chart.id }}"></div>
-        <chart-toolbar chart-type="mpld3" mpld3-chart-data="chart" chart-id="{{ chart.id }}">
-        </chart-toolbar>
+        <div mpld3-chart="chart" chart-id="{{ chart.id }}"></div>
       </div>
     </div>
   </div>
